+++ CHANGELOG.md
@@ -10,9 +10,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Answer responses for Forms
 - Language_code field on imports and exports
+- Remove whatsapp body hidden characters
 - Fix for assessment import for multiple languages
 ### Removed
-- Locale field on exports 
+- Locale field on exports
 -->
 
 ## v1.4.0 - 2024-12-18